+++ toolium/test/utils/test_dataset_map_param.py
@@ -21,11 +21,11 @@ import os
 import pytest
 
 from toolium.config_parser import ExtendedConfigParser
-from toolium.utils.dataset import map_param
-from toolium.utils.dataset import hide_passwords
+from toolium.utils import dataset
+from toolium.utils.dataset import map_param, set_base64_path, set_file_path, hide_passwords
 
 
-def test_a_env_param():
+def test_an_env_param():
     """
     Verification of a mapped parameter as ENV
     """
@@ -42,6 +42,9 @@ def test_a_file_param():
     result = map_param("[FILE:toolium/test/resources/document.txt]")
     expected = "Document used to verify functionalities in MSS "
     assert expected == result
+    set_file_path('toolium/test/resources/')
+    result = map_param("[FILE:document.txt]")
+    assert expected == result
 
 
 def test_a_base64_param():
@@ -51,16 +54,18 @@ def test_a_base64_param():
     result = map_param("[BASE64:toolium/test/resources/document.txt]")
     expected = "RG9jdW1lbnQgdXNlZCB0byB2ZXJpZnkgZnVuY3Rpb25hbGl0aWVzIGluIE1TUyA="
     assert expected == result
+    set_base64_path('toolium/test/resources/')
+    result = map_param("[BASE64:document.txt]")
+    assert expected == result
 
 
 def test_a_lang_param():
     """
     Verification of a mapped parameter as LANG
     """
-    context = mock.MagicMock()
-    context.language_dict = {"home": {"button": {"send": {"es": "enviar", "en": "send"}}}}
-    context.language = "es"
-    result = map_param("[LANG:home.button.send]", context)
+    dataset.language_terms = {"home": {"button": {"send": {"es": "enviar", "en": "send"}}}}
+    dataset.language = "es"
+    result = map_param("[LANG:home.button.send]")
     expected = "enviar"
     assert expected == result
 
@@ -69,10 +74,9 @@ def test_a_toolium_param():
     """
     Verification of a mapped parameter as TOOLIUM
     """
-    context = mock.MagicMock()
     config_file_path = os.path.join("toolium", "test", "resources", "toolium.cfg")
-    context.toolium_config = ExtendedConfigParser.get_config_from_file(config_file_path)
-    result = map_param("[TOOLIUM:TestExecution_environment]", context)
+    dataset.toolium_config = ExtendedConfigParser.get_config_from_file(config_file_path)
+    result = map_param("[TOOLIUM:TestExecution_environment]")
     expected = "QA"
     assert expected == result
 
@@ -81,9 +85,8 @@ def test_a_conf_param():
     """
     Verification of a mapped parameter as CONF
     """
-    context = mock.MagicMock()
-    context.project_config = {"service": {"port": 80}}
-    result = map_param("[CONF:service.port]", context)
+    dataset.project_config = {"service": {"port": 80}}
+    result = map_param("[CONF:service.port]")
     expected = 80
     assert expected == result
 
@@ -95,15 +98,78 @@ def test_a_context_param():
     context = mock.MagicMock()
     context.attribute = "attribute value"
     context.storage = {"storage_key": "storage entry value"}
-    result_att = map_param("[CONTEXT:attribute]", context)
+    dataset.behave_context = context
+
+    result_att = map_param("[CONTEXT:attribute]")
     expected_att = "attribute value"
     assert expected_att == result_att
-    result_st = map_param("[CONTEXT:storage_key]", context)
+    result_st = map_param("[CONTEXT:storage_key]")
     expected_st = "storage entry value"
     assert expected_st == result_st
 
 
+def test_a_context_param_with_dots():
+    """
+    Verification of a mapped parameter with dots as CONTEXT
+    """
+    context = mock.MagicMock()
+
+    class Obj(object):
+        pass
+    one = Obj()
+    one.two = "the value"
+
+    context.one = one
+    dataset.behave_context = context
+
+    result_att = map_param("[CONTEXT:one.two]")
+    expected_att = "the value"
+    assert expected_att == result_att
+
+    three = Obj()
+    three.four = "the other value"
+
+    context.storage = {"three": three}
+    dataset.behave_context = context
+
+    result_att = map_param("[CONTEXT:three.four]")
+    expected_att = "the other value"
+    assert expected_att == result_att
+
+
+def test_a_context_param_deprecated():
+    """
+    Verification of a mapped parameter as CONTEXT and check that deprecated message is logged
+    """
+    context = mock.MagicMock()
+    context.attribute = "attribute value"
+    dataset.logger = mock.MagicMock()
+
+    result_att = map_param("[CONTEXT:attribute]", context)
+    expected_att = "attribute value"
+    assert expected_att == result_att
+    dataset.logger.warning.assert_called_with("Deprecated context parameter has been sent to map_param method. Please, "
+                                              "configure dataset global variables instead of passing context to "
+                                              "map_param.")
+
+
 def test_a_poe_param_single_result():
+    """
+    Verification of a POE mapped parameter with a single result for a reference
+    """
+    dataset.poeditor_terms = [
+        {
+            "term": "Poniendo mute",
+            "definition": "Ahora la tele está silenciada",
+            "reference": "home:home.tv.mute",
+        }
+    ]
+    result = map_param('[POE:home.tv.mute]')
+    expected = "Ahora la tele está silenciada"
+    assert result == expected
+
+
+def test_a_poe_param_single_result_deprecated():
     """
     Verification of a POE mapped parameter with a single result for a reference
     """
@@ -115,17 +181,21 @@ def test_a_poe_param_single_result():
             "reference": "home:home.tv.mute",
         }
     ]
+    dataset.logger = mock.MagicMock()
+
     result = map_param('[POE:home.tv.mute]', context)
     expected = "Ahora la tele está silenciada"
     assert result == expected
+    dataset.logger.warning.assert_called_with("Deprecated context parameter has been sent to map_param method. Please, "
+                                              "configure dataset global variables instead of passing context to "
+                                              "map_param.")
 
 
 def test_a_poe_param_no_result_assertion():
     """
     Verification of a POE mapped parameter without result
     """
-    context = mock.MagicMock()
-    context.poeditor_export = [
+    dataset.poeditor_terms = [
         {
             "term": "Poniendo mute",
             "definition": "Ahora la tele está silenciada",
@@ -133,7 +203,7 @@ def test_a_poe_param_no_result_assertion():
         }
     ]
     with pytest.raises(Exception) as excinfo:
-        map_param('[POE:home.tv.off]', context)
+        map_param('[POE:home.tv.off]')
     assert "No translations found in POEditor for reference home.tv.off" in str(excinfo.value)
 
 
@@ -141,9 +211,8 @@ def test_a_poe_param_prefix_with_no_definition():
     """
     Verification of a POE mapped parameter with a single result for a reference
     """
-    context = mock.MagicMock()
-    context.project_config = {'poeditor': {'key_field': 'reference', 'search_type': 'contains', 'prefixes': ['PRE.']}}
-    context.poeditor_export = [
+    dataset.project_config = {'poeditor': {'key_field': 'reference', 'search_type': 'contains', 'prefixes': ['PRE.']}}
+    dataset.poeditor_terms = [
         {
             "term": "Hola, estoy aquí para ayudarte",
             "definition": None,
@@ -155,7 +224,7 @@ def test_a_poe_param_prefix_with_no_definition():
             "reference": "common:common.greetings.main",
         }
     ]
-    result = map_param('[POE:common:common.greetings.main]', context)
+    result = map_param('[POE:common:common.greetings.main]')
     expected = "Hola, buenas"
     assert result == expected
 
@@ -164,9 +233,8 @@ def test_a_poe_param_single_result_selecting_a_key_field():
     """
     Verification of a POE mapped parameter with a single result for a term
     """
-    context = mock.MagicMock()
-    context.project_config = {'poeditor': {'key_field': 'term'}}
-    context.poeditor_export = [
+    dataset.project_config = {'poeditor': {'key_field': 'term'}}
+    dataset.poeditor_terms = [
         {
             "term": "loginSelectLine_text_subtitle",
             "definition": "Te damos la bienvenida",
@@ -176,7 +244,7 @@ def test_a_poe_param_single_result_selecting_a_key_field():
             "comment": ""
         }
     ]
-    result = map_param('[POE:loginSelectLine_text_subtitle]', context)
+    result = map_param('[POE:loginSelectLine_text_subtitle]')
     expected = "Te damos la bienvenida"
     assert result == expected
 
@@ -185,9 +253,8 @@ def test_a_poe_param_multiple_results():
     """
     Verification of a POE mapped parameter with several results for a reference
     """
-    context = mock.MagicMock()
-    context.project_config = {'poeditor': {'key_field': 'reference'}}
-    context.poeditor_export = [
+    dataset.project_config = {'poeditor': {'key_field': 'reference'}}
+    dataset.poeditor_terms = [
         {
             "term": "Hola, estoy aquí para ayudarte",
             "definition": "Hola, estoy aquí para ayudarte",
@@ -199,7 +266,7 @@ def test_a_poe_param_multiple_results():
             "reference": "common:common.greetings.main",
         }
     ]
-    result = map_param('[POE:common.greetings.main]', context)
+    result = map_param('[POE:common.greetings.main]')
     first_expected = "Hola, estoy aquí para ayudarte"
     second_expected = "Hola, buenas"
     assert len(result) == 2
@@ -211,9 +278,8 @@ def test_a_poe_param_multiple_options_but_only_one_result():
     """
     Verification of a POE mapped parameter with a single result from several options for a key
     """
-    context = mock.MagicMock()
-    context.project_config = {'poeditor': {'key_field': 'term', 'search_type': 'exact'}}
-    context.poeditor_export = [
+    dataset.project_config = {'poeditor': {'key_field': 'term', 'search_type': 'exact'}}
+    dataset.poeditor_terms = [
         {
             "term": "loginSelectLine_text_subtitle",
             "definition": "Te damos la bienvenida_1",
@@ -231,7 +297,7 @@ def test_a_poe_param_multiple_options_but_only_one_result():
             "comment": ""
         }
     ]
-    result = map_param('[POE:loginSelectLine_text_subtitle]', context)
+    result = map_param('[POE:loginSelectLine_text_subtitle]')
     expected = "Te damos la bienvenida_1"
     assert result == expected
 
@@ -240,9 +306,8 @@ def test_a_poe_param_with_prefix():
     """
     Verification of a POE mapped parameter with several results for a reference, filtered with a prefix
     """
-    context = mock.MagicMock()
-    context.project_config = {'poeditor': {'key_field': 'reference', 'search_type': 'contains', 'prefixes': ['PRE.']}}
-    context.poeditor_export = [
+    dataset.project_config = {'poeditor': {'key_field': 'reference', 'search_type': 'contains', 'prefixes': ['PRE.']}}
+    dataset.poeditor_terms = [
         {
             "term": "Hola, estoy aquí para ayudarte",
             "definition": "Hola, estoy aquí para ayudarte",
@@ -254,7 +319,7 @@ def test_a_poe_param_with_prefix():
             "reference": "common:PRE.common.greetings.main",
         }
     ]
-    result = map_param('[POE:common.greetings.main]', context)
+    result = map_param('[POE:common.greetings.main]')
     expected = "Hola, buenas"
     assert result == expected
 
@@ -263,9 +328,8 @@ def test_a_poe_param_with_two_prefixes():
     """
     Verification of a POE mapped parameter with several results for a reference, filtered with two prefixes
     """
-    context = mock.MagicMock()
-    context.project_config = {'poeditor': {'prefixes': ['MH.', 'PRE.']}}
-    context.poeditor_export = [
+    dataset.project_config = {'poeditor': {'prefixes': ['MH.', 'PRE.']}}
+    dataset.poeditor_terms = [
         {
             "term": "Hola, estoy aquí para ayudarte",
             "definition": "Hola, estoy aquí para ayudarte",
@@ -282,7 +346,7 @@ def test_a_poe_param_with_two_prefixes():
             "reference": "common:MH.common.greetings.main",
         }
     ]
-    result = map_param('[POE:common.greetings.main]', context)
+    result = map_param('[POE:common.greetings.main]')
     expected = "Hola, buenas MH"
     assert result == expected
 
@@ -291,9 +355,8 @@ def test_a_poe_param_with_prefix_and_exact_resource():
     """
     Verification of a POE mapped parameter that uses an exact resource name and has a prefix configured
     """
-    context = mock.MagicMock()
-    context.project_config = {'poeditor': {'prefixes': ['PRE.']}}
-    context.poeditor_export = [
+    dataset.project_config = {'poeditor': {'prefixes': ['PRE.']}}
+    dataset.poeditor_terms = [
         {
             "term": "Hola, estoy aquí para ayudarte",
             "definition": "Hola, estoy aquí para ayudarte",
@@ -305,7 +368,7 @@ def test_a_poe_param_with_prefix_and_exact_resource():
             "reference": "common:PRE.common.greetings.main",
         }
     ]
-    result = map_param('[POE:common:common.greetings.main]', context)
+    result = map_param('[POE:common:common.greetings.main]')
     expected = "Hola, buenas"
     assert result == expected
 
@@ -333,9 +396,8 @@ def test_a_combi_of_textplusconfig_integer():
     """
     Verification of a combination of text plus a config param
     """
-    context = mock.MagicMock()
-    context.project_config = {"service": {"port": 80}}
-    result = map_param("use port [CONF:service.port]", context)
+    dataset.project_config = {"service": {"port": 80}}
+    result = map_param("use port [CONF:service.port]")
     expected = "use port 80"
     assert expected == result
 
@@ -362,13 +424,36 @@ def test_a_combi_of_config_plustext_plusconfig():
     assert expected == result
 
 
+def test_a_combi_of_config_inside_config():
+    """
+    Verification of a combination of a config param inside another config param
+    """
+    dataset.project_config = {"var_name": "NAME2"}
+    os.environ['MY_VAR_NAME1'] = "name1 value"
+    os.environ['MY_VAR_NAME2'] = "name2 value"
+    result = map_param("[ENV:MY_VAR_[CONF:var_name]]")
+    expected = "name2 value"
+    assert expected == result
+
+
+def test_a_combi_of_config_inside_config_recursively():
+    """
+    Verification of a combination of a config param inside another config param recursively
+    """
+    dataset.project_config = {"var_number": 2, "var_name_1": "A", "var_name_2": "B"}
+    os.environ['MY_VAR_A'] = "A value"
+    os.environ['MY_VAR_B'] = "B value"
+    result = map_param("[CONF:var_number] some text [ENV:MY_VAR_[CONF:var_name_[CONF:var_number]]]")
+    expected = "2 some text B value"
+    assert expected == result
+
+
 def test_a_conf_param_with_special_characters():
     """
     Verification of a combination of text plus a config param with special characters
     """
-    context = mock.MagicMock()
-    context.project_config = {"user": "user-1", "password": "p4:ssw0_rd"}
-    result = map_param("[CONF:user]-an:d_![CONF:password]", context)
+    dataset.project_config = {"user": "user-1", "password": "p4:ssw0_rd"}
+    result = map_param("[CONF:user]-an:d_![CONF:password]")
     expected = "user-1-an:d_!p4:ssw0_rd"
     assert expected == result