+++ toolium/test/utils/test_dataset_replace_param.py
@@ -17,6 +17,7 @@ limitations under the License.
 """
 
 import datetime
+from uuid import UUID
 
 from toolium.utils.dataset import replace_param
 
@@ -106,6 +107,11 @@ def test_replace_param_blank():
     assert param == ' '
 
 
+def test_replace_param_uuid():
+    param = replace_param('[UUID]')
+    assert UUID(param).version == 4
+
+
 def test_replace_param_random():
     param = replace_param('[RANDOM]')
     assert len(param) == 8
@@ -188,10 +194,133 @@ def test_replace_param_now_not_spanish():
     assert param == datetime.datetime.utcnow().strftime('%Y/%m/%d %H:%M:%S')
 
 
+def test_replace_param_now_with_format():
+    param = replace_param('[NOW(%Y-%m-%dT%H:%M:%SZ)]')
+    assert param == datetime.datetime.utcnow().strftime('%Y-%m-%dT%H:%M:%SZ')
+
+
+def test_not_replace_param_now_with_invalid_opening_parenthesis_in_format():
+    param = replace_param('[NOW(%Y-%m-%dT(%H:%M:%SZ)]')
+    assert param == '[NOW(%Y-%m-%dT(%H:%M:%SZ)]'
+
+
+def test_not_replace_param_now_with_invalid_closing_parenthesis_in_format():
+    param = replace_param('[NOW(%Y-%m-%dT)%H:%M:%SZ)]')
+    assert param == '[NOW(%Y-%m-%dT)%H:%M:%SZ)]'
+
+
 def test_replace_param_now_offset():
     param = replace_param('[NOW + 5 MINUTES]', language='es')
     assert param == datetime.datetime.strftime(
         datetime.datetime.utcnow() + datetime.timedelta(minutes=5), '%d/%m/%Y %H:%M:%S')
+
+
+def test_replace_param_now_offset_with_format():
+    param = replace_param('[NOW(%Y-%m-%dT%H:%M:%SZ) + 5 MINUTES]')
+    assert param == datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=5), '%Y-%m-%dT%H:%M:%SZ')
+
+
+def test_replace_param_today_offset_and_more():
+    param = replace_param('The day [TODAY - 1 DAYS] was yesterday', language='es')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.today() - datetime.timedelta(days=1), '%d/%m/%Y')
+    assert param == f'The day {offset_date} was yesterday'
+
+
+def test_replace_param_today_offset_and_more_not_spanish():
+    param = replace_param('The day [TODAY - 1 DAYS] was yesterday', language='it')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.today() - datetime.timedelta(days=1), '%Y/%m/%d')
+    assert param == f'The day {offset_date} was yesterday'
+
+
+def test_replace_param_now_offset_and_more():
+    param = replace_param('I will arrive at [NOW + 10 MINUTES] tomorrow', language='es')
+    offset_datetime = datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=10), '%d/%m/%Y %H:%M:%S')
+    assert param == f'I will arrive at {offset_datetime} tomorrow'
+
+
+def test_replace_param_now_offset_and_more_not_spanish():
+    param = replace_param('I will arrive at [NOW + 10 MINUTES] tomorrow', language='it')
+    offset_datetime = datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=10), '%Y/%m/%d %H:%M:%S')
+    assert param == f'I will arrive at {offset_datetime} tomorrow'
+
+
+def test_replace_param_now_offset_with_format_and_more():
+    param = replace_param('I will arrive at [NOW(%Y-%m-%dT%H:%M:%SZ) + 10 MINUTES] tomorrow')
+    offset_datetime = datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=10), '%Y-%m-%dT%H:%M:%SZ')
+    assert param == f'I will arrive at {offset_datetime} tomorrow'
+
+
+def test_replace_param_now_offset_with_format_and_more_language_is_irrelevant():
+    param = replace_param('I will arrive at [NOW(%Y-%m-%dT%H:%M:%SZ) + 10 MINUTES] tomorrow', language='ru')
+    offset_datetime = datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=10), '%Y-%m-%dT%H:%M:%SZ')
+    assert param == f'I will arrive at {offset_datetime} tomorrow'
+
+
+def test_replace_param_today_offset_with_format_and_more_with_extra_spaces():
+    param = replace_param('The date [NOW(%Y-%m-%dT%H:%M:%SZ)   - 1    DAYS ] was yesterday')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.utcnow() - datetime.timedelta(days=1), '%Y-%m-%dT%H:%M:%SZ')
+    assert param == f'The date {offset_date} was yesterday'
+
+
+def test_replace_param_today_offset_and_more_with_extra_spaces():
+    param = replace_param('The day [TODAY    - 1    DAYS ] was yesterday', language='es')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.today() - datetime.timedelta(days=1), '%d/%m/%Y')
+    assert param == f'The day {offset_date} was yesterday'
+
+
+def test_replace_param_today_offset_and_more_at_the_end():
+    param = replace_param('Yesterday was [TODAY - 1 DAYS]', language='es')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.today() - datetime.timedelta(days=1), '%d/%m/%Y')
+    assert param == f'Yesterday was {offset_date}'
+
+
+def test_replace_param_today_offset_and_more_at_the_beginning():
+    param = replace_param('[TODAY - 1 DAYS] is yesterday', language='es')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.today() - datetime.timedelta(days=1), '%d/%m/%Y')
+    assert param == f'{offset_date} is yesterday'
+
+
+def test_replace_param_today_offsets_and_more():
+    param = replace_param(
+        'The day [TODAY - 1 DAYS] was yesterday and I have an appointment at [NOW + 10 MINUTES]', language='es')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.today() - datetime.timedelta(days=1), '%d/%m/%Y')
+    offset_datetime = datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=10), '%d/%m/%Y %H:%M:%S')
+    assert param == f'The day {offset_date} was yesterday and I have an appointment at {offset_datetime}'
+
+
+def test_replace_param_now_offsets_with_format_and_more():
+    param = replace_param(
+        'The date [NOW(%Y-%m-%dT%H:%M:%SZ) - 1 DAYS] was yesterday '
+        'and I have an appointment at [NOW(%Y-%m-%dT%H:%M:%SZ) + 10 MINUTES]')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.utcnow() - datetime.timedelta(days=1), '%Y-%m-%dT%H:%M:%SZ')
+    offset_datetime = datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=10), '%Y-%m-%dT%H:%M:%SZ')
+    assert param == f'The date {offset_date} was yesterday and I have an appointment at {offset_datetime}'
+
+
+def test_replace_param_now_offsets_with_and_without_format_and_more():
+    param = replace_param(
+        'The date [NOW(%Y-%m-%dT%H:%M:%SZ) - 1 DAYS] was yesterday '
+        'and I have an appointment at [NOW + 10 MINUTES]', language='es')
+    offset_date = datetime.datetime.strftime(
+        datetime.datetime.utcnow() - datetime.timedelta(days=1), '%Y-%m-%dT%H:%M:%SZ')
+    offset_datetime = datetime.datetime.strftime(
+        datetime.datetime.utcnow() + datetime.timedelta(minutes=10), '%d/%m/%Y %H:%M:%S')
+    assert param == f'The date {offset_date} was yesterday and I have an appointment at {offset_datetime}'
 
 
 def test_replace_param_str_int():