+++ toolium/visual_test.py
@@ -24,7 +24,6 @@ import re
 import shutil
 from io import BytesIO
 from os import path
-
 from selenium.common.exceptions import NoSuchElementException
 
 from toolium.driver_wrappers_pool import DriverWrappersPool
@@ -152,6 +151,7 @@ class VisualTest(object):
         img = Image.open(BytesIO(self.driver_wrapper.driver.get_screenshot_as_png()))
         img = self.remove_scrolls(img)
         img = self.mobile_resize(img)
+        img = self.desktop_resize(img)
         img = self.exclude_elements(img, exclude_web_elements)
         img = self.crop_element(img, web_element)
         img.save(output_file)
@@ -215,10 +215,29 @@ class VisualTest(object):
         :returns: modified image object
         """
         if self.driver_wrapper.is_ios_test() or self.driver_wrapper.is_android_web_test():
-            scale = img.size[0] / self.utils.get_window_size()['width']
-            if scale != 1:
-                new_image_size = (int(img.size[0] / scale), int(img.size[1] / scale))
-                img = img.resize(new_image_size, Image.ANTIALIAS)
+            img = self.base_resize(img=img)
+        return img
+
+    def desktop_resize(self, img):
+        """Resize image in Mac with Retina to fit window size
+
+        :param img: image object
+        :returns: modified image object
+        """
+        if self.driver_wrapper.is_mac_test():
+            img = self.base_resize(img=img)
+        return img
+
+    def base_resize(self, img):
+        """Base method for resize image
+
+        :param img: image object
+        :returns: modified image object
+        """
+        scale = img.size[0] / self.utils.get_window_size()['width']
+        if scale != 1:
+            new_image_size = (int(img.size[0] / scale), int(img.size[1] / scale))
+            img = img.resize(new_image_size, Image.ANTIALIAS)
         return img
 
     def get_element_box(self, web_element):
@@ -321,7 +340,11 @@ class VisualTest(object):
         :param message: error message
         """
         self.results[result] += 1
-        row = self._get_html_row(result, report_name, image_file, baseline_file, message)
+        output_baseline_file = None
+        if baseline_file is not None:
+            output_baseline_file = os.path.join(self.output_directory, os.path.basename(baseline_file))
+            shutil.copyfile(baseline_file, output_baseline_file)
+        row = self._get_html_row(result, report_name, image_file, output_baseline_file, message)
         self._add_data_to_report_before_tag(row, '</tbody>')
         self._update_report_summary()
 
@@ -427,3 +450,14 @@ class VisualTest(object):
             diff_message = message
 
         return diff_message
+
+    @staticmethod
+    def update_latest_report():
+        """
+        Duplicate current visual testing report into latest folder
+        """
+        if os.path.exists(DriverWrappersPool.visual_output_directory):
+            latest_directory = os.path.join(os.path.dirname(DriverWrappersPool.visual_output_directory), 'latest')
+            if os.path.exists(latest_directory):
+                shutil.rmtree(latest_directory)
+            shutil.copytree(DriverWrappersPool.visual_output_directory, latest_directory)